--- packages/api-reference/src/features/ExampleRequest/index.ts
@@ -1,2 +1,0 @@
-export { default as ExamplePicker } from './ExamplePicker.vue'
-export { default as ExampleRequest } from './ExampleRequest.vue'
