+++ packages/api-reference/src/features/RequestExample/index.ts
@@ -1,0 +1,2 @@
+export { default as ExamplePicker } from './ExamplePicker.vue'
+export { default as RequestExample } from './RequestExample.vue'
